requirements.txt
Flask==0.12
gunicorn==19.6.0
GroupyAPI==0.8.1